requirements.txt
qrcode
streamlit_option_menu
pytz
python-dotenv
pymongo